pyproject.toml
[project]
name = "fnos-git-auth"
version = "0.1.0"
description = "飞牛 fnOS Git 认证工具，自动配置 git extraHeader 实现免密访问 NAS Git 仓库"
readme = "README.md"
authors = [
    { name = "donnel", email = "donnel.lu@foxmail.com" }
]
requires-python = ">=3.10"
keywords = ["fnos", "git", "authentication", "nas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "websockets>=12.0",
    "pycryptodome>=3.20",
    "click>=8.1",
]

[project.scripts]
fnos-git-auth = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
